aYourShoes/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
cx-Oracle==8.2.1
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
marshmallow==2.15.6
Pillow==8.3.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
transbank-sdk==1.5.0
urllib3==1.26.6
